Portfolio Project Revisions - CT1: Updated Pseudocode
I noticed that I received a “Distinguished,” grade on my pseudocode for Critical Thinking Module 1. However, I was still docked a few points for my pseudocode being sub-par, so I decided to correct that and make it a bit better. Prior to this course I had only ever really read Pseudocode and not really wrote it myself, a few simple outlines but nothing to the extent of this class. I feel as though since this assignment I have focused more on my pseudocode and have created better pseudocodes since then.